back/requirements.txt
fastapi
uvicorn
python-dotenv
cohere
flask
flask-cors
requests
pymongo
dnspython
